feat(react-components): Move global camera controls to Reveal context
